bot/pyproject.toml
[project]
name = "test-task-tg-profile"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "aiogram>=3.18.0",
    "fastapi>=0.115.11",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.0.1",
    "tortoise-orm>=0.24.2",
    "uvicorn>=0.34.0",
]